azure-ai-agents/requirements.txt
azure-ai-agents==1.1.0b2
aiohttp==3.12.12
azure-ai-projects==1.0.0b11
azure-identity==1.23.0
azure-mgmt-logic==10.0.0
azure-mgmt-core==1.5.0
azure-common==1.1.28
dotenv==1.1.0